topic mapper null value handling

+++ backend/src/main/java/com/proj/forum/dto/TopicDto.java
@@ -29,6 +29,6 @@ public record TopicDto(
         Double userRate,
         int userRateCount,
         List<CommentDto> comments,
-        int myRate
+        Short myRate
 ) {
 }
